requirements.txt
streamlit
supabase
pandas
pdfplumber
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
plotly
